Refactor numerical_grids_and_timestep to allow parallelization extensions
